modified:   setup.sh Addresses problems creating Virtual machines through the use of specific     options better targetted for a now pre-determinated location.
Replace derived Resource group with one to be created.
Replace derived location with determination of 'eastus'.
Added creation of Resource group at location.
Added prompt to for user to confirm awareness of location.
Added courtesy notice concerning Cleanup.

As the script now takes some further measure of repsonsibility for the
    SizeSKU of the created machines, added commenting regarding cost
    awareness and how to select other SizeSKUs.